psnr-pae works, and script can test encode-decode

+++ tests/psnr_pae.cpp
@@ -314,7 +314,7 @@ void find_psnr_pae(const img_info& img1, const img_info& img2,
   float t = 0;
   for (ui32 c = 0; c < img1.num_comps; ++c)
     t += (float)mse[c];
-  t /= num_pixels;
+  t /= (float)num_pixels;
   psnr = 10.0f * log10f((float)img1.max_val * (float)img1.max_val / t);
 }
